fix deps and migration
