+++ .devcontainer/.dev_config.yaml
@@ -1,3 +1,4 @@
+rabbitmq_host: rabbitmq
 db_url: postgresql://admin:admin@postgresql/inbox
 db_print_logs: false
 s3_endpoint_url: http://localstack:4566
