--- db_migration/alembic/versions/20220126_174604_b0f0c67ac176_add_enum_uploadstate_to_table.py
@@ -1,36 +1,0 @@
-"""Add enum 'UploadState' to table
-
-Revision ID: b0f0c67ac176
-Revises: 19e3d538a7cb
-Create Date: 2022-01-26 17:46:04.539675
-
-"""
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = "b0f0c67ac176"
-down_revision = "19e3d538a7cb"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "files",
-        sa.Column(
-            "state",
-            sa.Enum(
-                "REGISTERED", "PENDING", "CONFIRMED", "COMPLETED", name="uploadstate"
-            ),
-            nullable=False,
-        ),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("files", "state")
-    # ### end Alembic commands ###
