+++ setup.cfg
@@ -35,7 +35,7 @@ zip_safe = False
 include_package_data = True
 packages = find:
 install_requires =
-    ghga-service-chassis-lib[api,pubsub,postgresql,s3]==0.10.0
+    ghga-service-chassis-lib[api,pubsub,postgresql,s3]==0.11.0
     ghga-message-schemas==0.3.1
 
 python_requires = >= 3.9
@@ -46,7 +46,7 @@ console_scripts =
 
 [options.extras_require]
 dev =
-    ghga-service-chassis-lib[dev]==0.10.0
+    ghga-service-chassis-lib[dev]==0.11.0
     postgres.client
 db_migration =
     alembic==1.6.5
